reorganise `types` module to introduce namespacing
