+++ src/sync/critical_section.rs
@@ -41,7 +41,7 @@
 //! possibly surprising ways.
 
 #[cfg(all(Py_3_14, not(Py_LIMITED_API)))]
-use crate::types::PyMutex;
+use crate::sync::PyMutex;
 
 #[cfg(all(Py_3_14, not(Py_LIMITED_API)))]
 use crate::Python;
@@ -276,7 +276,7 @@ mod tests {
     #[allow(unused_imports, reason = "conditionally used")]
     use crate::platform::prelude::*;
     #[cfg(all(not(Py_LIMITED_API), Py_3_14))]
-    use crate::types::PyMutex;
+    use crate::sync::PyMutex;
     #[cfg(feature = "macros")]
     use core::sync::atomic::{AtomicBool, Ordering};
     #[cfg(any(feature = "macros", all(not(Py_LIMITED_API), Py_3_14)))]
