+++ src/types/dict.rs
@@ -1,3 +1,5 @@
+//! Python dictionaries and related types.
+
 use crate::err::{self, PyErr, PyResult};
 use crate::ffi::Py_ssize_t;
 use crate::ffi_ptr_ext::FfiPtrExt;
@@ -11,6 +13,10 @@ use crate::{
     types::{PyType, PyTypeMethods},
     Py,
 };
+
+pub(crate) mod items;
+pub(crate) mod keys;
+pub(crate) mod values;
 
 /// Represents a Python `dict`.
 ///
@@ -45,48 +51,6 @@ pyobject_native_type_core!(
     "builtins",
     "dict",
     #checkfunction=ffi::PyDict_Check
-);
-
-/// Represents a Python `dict_keys`.
-#[cfg(not(any(PyPy, GraalPy, RustPython)))]
-#[repr(transparent)]
-pub struct PyDictKeys(PyAny);
-
-#[cfg(not(any(PyPy, GraalPy, RustPython)))]
-pyobject_native_type_core!(
-    PyDictKeys,
-    pyobject_native_static_type_object!(ffi::PyDictKeys_Type),
-    "builtins",
-    "dict_keys",
-    #checkfunction=ffi::PyDictKeys_Check
-);
-
-/// Represents a Python `dict_values`.
-#[cfg(not(any(PyPy, GraalPy, RustPython)))]
-#[repr(transparent)]
-pub struct PyDictValues(PyAny);
-
-#[cfg(not(any(PyPy, GraalPy, RustPython)))]
-pyobject_native_type_core!(
-    PyDictValues,
-    pyobject_native_static_type_object!(ffi::PyDictValues_Type),
-    "builtins",
-    "dict_values",
-    #checkfunction=ffi::PyDictValues_Check
-);
-
-/// Represents a Python `dict_items`.
-#[cfg(not(any(PyPy, GraalPy, RustPython)))]
-#[repr(transparent)]
-pub struct PyDictItems(PyAny);
-
-#[cfg(not(any(PyPy, GraalPy, RustPython)))]
-pyobject_native_type_core!(
-    PyDictItems,
-    pyobject_native_static_type_object!(ffi::PyDictItems_Type),
-    "builtins",
-    "dict_items",
-    #checkfunction=ffi::PyDictItems_Check
 );
 
 impl PyDict {
@@ -954,6 +918,8 @@ mod tests {
     use crate::platform::prelude::*;
     use crate::platform::HashMap;
     use crate::types::{PyAnyMethods as _, PyTuple};
+    #[cfg(not(any(PyPy, GraalPy, RustPython)))]
+    use crate::types::{PyDictItems, PyDictKeys, PyDictValues};
     use alloc::collections::BTreeMap;
 
     #[test]
