+++ src/types/frozendict.rs
@@ -1,3 +1,5 @@
+//! Python frozen dictionaries and related types.
+
 use crate::err::{PyErr, PyResult};
 use crate::ffi_ptr_ext::FfiPtrExt;
 use crate::instance::Bound;
