+++ src/types/frozenset.rs
@@ -1,3 +1,5 @@
+//! Python frozen sets and related types.
+
 use crate::types::PyIterator;
 use crate::{
     err::{self, PyErr, PyResult},
